+++ logging/bin/remove_openshift_routes.sh
@@ -10,6 +10,13 @@ this_script=$(basename "$0")
 
 log_debug "Script [$this_script] has started [$(date)]"
 
+if [ "$OPENSHIFT_CLUSTER" != "true" ]; then
+    if [ "${CHECK_OPENSHIFT_CLUSTER:-true}" == "true" ]; then
+        log_error "This script should only be run on OpenShift clusters"
+        exit 1
+    fi
+fi
+
 oc -n "$LOG_NS" delete route v4m-es-kibana-svc --ignore-not-found
 oc -n "$LOG_NS" delete route v4m-es-client-service --ignore-not-found
 oc -n "$LOG_NS" delete route v4m-search --ignore-not-found